+++ test/formatters/cli.js
@@ -3,24 +3,24 @@ var buster = require('buster'),
 
 buster.testCase('cli - format', {
   'should format cli line properly when desc exists': function () {
-    var result = cli.format([{ uri: 'http://somehost', status: 'ok', successes: ['somesuccess'], duration: 20 }]);
+    var result = cli.format([{ uri: 'http://somehost', status: 'success', successes: ['somesuccess'], duration: 20 }]);
     assert.equals(result, [
-      'ok'.green + ' - ' + 'http://somehost'.cyan + ' - 20ms',
+      'success'.green + ' - ' + 'http://somehost'.cyan + ' - 20ms',
       ' * somesuccess',
       ''
     ]);
   },
   'should format cli line properly when name exists': function () {
-    var result = cli.format([{ uri: 'http://somehost', status: 'ok', name: 'someapp', duration: 20 }]);
+    var result = cli.format([{ uri: 'http://somehost', status: 'success', name: 'someapp', duration: 20 }]);
     assert.equals(result, [
-      'ok'.green + ' - someapp - ' + 'http://somehost'.cyan + ' - 20ms',
+      'success'.green + ' - someapp - ' + 'http://somehost'.cyan + ' - 20ms',
       ''
     ]);
   },
   'should colourise status and uri and exclude name and desc from cli line when they do not exist': function () {
-    var result = cli.format([{ uri: 'http://somehost', status: 'ok', duration: 20 }]);
+    var result = cli.format([{ uri: 'http://somehost', status: 'success', duration: 20 }]);
     assert.equals(result, [
-      'ok'.green + ' - ' + 'http://somehost'.cyan + ' - 20ms',
+      'success'.green + ' - ' + 'http://somehost'.cyan + ' - 20ms',
       ''
     ]);
   },